made EditTools helper and Tools folder
