+++ app/src/components/map/Tools/DisplayPosition.tsx
